requirements.txt
streamlit
requests
python-dotenv
openai
transformers
torch
emoji